move all worker related files into a special worker directory
