+++ arangod/Pregel/PregelFeature.cpp
@@ -53,7 +53,7 @@
 #include "Pregel/ExecutionNumber.h"
 #include "Pregel/Recovery.h"
 #include "Pregel/Utils.h"
-#include "Pregel/Worker.h"
+#include "Pregel/Worker/Worker.h"
 #include "Pregel/WorkerConductorMessages.h"
 #include "RestServer/DatabasePathFeature.h"
 #include "Scheduler/Scheduler.h"
