+++ arangod/Pregel/Worker/Worker.cpp
@@ -34,7 +34,7 @@
 #include "Pregel/PregelFeature.h"
 #include "Pregel/Status/Status.h"
 #include "Pregel/VertexComputation.h"
-#include "Pregel/Worker.h"
+#include "Worker.h"
 
 #include "ApplicationFeatures/ApplicationServer.h"
 #include "Basics/WriteLocker.h"
